+++ packages/ketcher-polymer-editor-react/src/helpers/getDefaultPresets.test.tsx
@@ -17,202 +17,27 @@
 
 import { getDefaultPresets } from './getDefaultPreset';
 import { MonomerItemType } from 'components/monomerLibrary/monomerLibraryItem/types';
-import { Struct } from 'ketcher-core';
 import { IRnaPreset } from 'components/monomerLibrary/RnaBuilder/types';
+import {
+  monomers,
+  phosphate,
+  ribose,
+  cytosine,
+  guanine,
+  thymine,
+  uracil,
+  adenine,
+} from '../testMockData/monomerPresets';
 
 describe('getDefaultPreset function', () => {
   it('should return empty array if cannot return default nucteotides', () => {
-    const monomers: MonomerItemType[] = [
-      {
-        props: {
-          MonomerName: '12ddR',
-          Name: "1',2'-Di-Deoxy-Ribose",
-          MonomerNaturalAnalogCode: '',
-          MonomerType: '',
-          BranchMonomer: '',
-          MonomerCaps: '',
-          MonomerCode: '',
-        },
-        label: '',
-        struct: new Struct(),
-      },
-      {
-        props: {
-          MonomerName: '3A6',
-          Name: "6-amino-hexanol (3' end)",
-          MonomerNaturalAnalogCode: '',
-          MonomerType: '',
-          BranchMonomer: '',
-          MonomerCaps: '',
-          MonomerCode: '',
-        },
-        label: '',
-        struct: new Struct(),
-      },
-      {
-        props: {
-          MonomerName: '3FAM',
-          Name: '3-FAM',
-          MonomerNaturalAnalogCode: '',
-          MonomerType: '',
-          BranchMonomer: '',
-          MonomerCaps: '',
-          MonomerCode: '',
-        },
-        label: '',
-        struct: new Struct(),
-      },
-      {
-        props: {
-          MonomerName: '4sR',
-          Name: '4-Thio-Ribose',
-          MonomerNaturalAnalogCode: '',
-          MonomerType: '',
-          BranchMonomer: '',
-          MonomerCaps: '',
-          MonomerCode: '',
-        },
-        label: '',
-        struct: new Struct(),
-      },
-    ];
-
     const testArr = getDefaultPresets(monomers);
     expect(testArr).toBeEmptyArray();
   });
 
   it('should return array of default presets from monomers', () => {
-    const phosphate = {
-      props: {
-        MonomerName: 'P',
-        Name: 'Phosphate',
-        MonomerNaturalAnalogCode: '',
-        MonomerType: '',
-        BranchMonomer: '',
-        MonomerCaps: '',
-        MonomerCode: '',
-      },
-      label: '',
-      struct: new Struct(),
-    };
-    const ribose = {
-      props: {
-        MonomerName: 'R',
-        Name: 'Ribose',
-        MonomerNaturalAnalogCode: '',
-        MonomerType: '',
-        BranchMonomer: '',
-        MonomerCaps: '',
-        MonomerCode: '',
-      },
-      label: '',
-      struct: new Struct(),
-    };
-    const thymine = {
-      props: {
-        MonomerName: 'T',
-        Name: 'Thymine',
-        MonomerNaturalAnalogCode: '',
-        MonomerType: '',
-        BranchMonomer: '',
-        MonomerCaps: '',
-        MonomerCode: '',
-      },
-      label: '',
-      struct: new Struct(),
-    };
-    const cytosine = {
-      props: {
-        MonomerName: 'C',
-        Name: 'Cytosine',
-        MonomerNaturalAnalogCode: '',
-        MonomerType: '',
-        BranchMonomer: '',
-        MonomerCaps: '',
-        MonomerCode: '',
-      },
-      label: '',
-      struct: new Struct(),
-    };
-    const uracil = {
-      props: {
-        MonomerName: 'U',
-        Name: 'Uracil',
-        MonomerNaturalAnalogCode: '',
-        MonomerType: '',
-        BranchMonomer: '',
-        MonomerCaps: '',
-        MonomerCode: '',
-      },
-      label: '',
-      struct: new Struct(),
-    };
-    const adenine = {
-      props: {
-        MonomerName: 'A',
-        Name: 'Adenine',
-        MonomerNaturalAnalogCode: '',
-        MonomerType: '',
-        BranchMonomer: '',
-        MonomerCaps: '',
-        MonomerCode: '',
-      },
-      label: '',
-      struct: new Struct(),
-    };
-    const guanine = {
-      props: {
-        MonomerName: 'G',
-        Name: 'Guanine',
-        MonomerNaturalAnalogCode: '',
-        MonomerType: '',
-        BranchMonomer: '',
-        MonomerCaps: '',
-        MonomerCode: '',
-      },
-      label: '',
-      struct: new Struct(),
-    };
-    const monomers: MonomerItemType[] = [
-      {
-        props: {
-          MonomerName: '12ddR',
-          Name: "1',2'-Di-Deoxy-Ribose",
-          MonomerNaturalAnalogCode: '',
-          MonomerType: '',
-          BranchMonomer: '',
-          MonomerCaps: '',
-          MonomerCode: '',
-        },
-        label: '',
-        struct: new Struct(),
-      },
-      {
-        props: {
-          MonomerName: '3A6',
-          Name: "6-amino-hexanol (3' end)",
-          MonomerNaturalAnalogCode: '',
-          MonomerType: '',
-          BranchMonomer: '',
-          MonomerCaps: '',
-          MonomerCode: '',
-        },
-        label: '',
-        struct: new Struct(),
-      },
-      {
-        props: {
-          MonomerName: '3FAM',
-          Name: '3-FAM',
-          MonomerNaturalAnalogCode: '',
-          MonomerType: '',
-          BranchMonomer: '',
-          MonomerCaps: '',
-          MonomerCode: '',
-        },
-        label: '',
-        struct: new Struct(),
-      },
+    const monomerData: MonomerItemType[] = [
+      ...monomers,
       phosphate,
       ribose,
       cytosine,
@@ -236,7 +61,7 @@ describe('getDefaultPreset function', () => {
       phosphate,
     };
 
-    const testArr = getDefaultPresets(monomers);
+    const testArr = getDefaultPresets(monomerData);
     expect(testArr).toContainEqual(thymineNucleotide);
     expect(testArr).toContainEqual(guanineNucleotide);
     expect(testArr).toContainEqual(
